The HLSL file should preserve the original line numbers
This version 1.7.31

Added a new class NewLineCounterStream that can wrap a ostream instance.
All strings that are streamed into NewLineCounterStream are scanned for
the '\n' (new line) character. Each '\n' character found is counted.

With the help of NewLineCounterStream, which is instantiated as
CodeEmitter::m_out, We can appoximate the output line number of the
HLSL symbols to be as close as possible to the input line number from
the AZSL file for the corresponding symbols.

The line numbers are simply matched by adding new lines to the
output file only if the current line number of the symbol about to be
output is less than the original line number.

The idea is that later if there are errors in the HLSL file, when
compiled with other tools like DXC, the error line number will be very
close (+/-) 2 lines to the original line number.

Signed-off-by: garrieta <garrieta@amazon.com>