Removed the --use-unbounded-spaces command line option and made this feature always-on insetead.
While testing I realized the the unbounded arrays validation was overly aggressive. It was applying DirectX restrictions to all platforms, even though DirectX is the only platform where unbounded arrays consume all available registers. After discussing with @jeremyong-az and @moudgils it was determined that using a unique space per unbounded array would not be detrimental to the other platforms, because they don't have any concept of register space anyway, and the spaces are ignored. So rather than support two modes of operation, we just apply unique per-unbounded-array spaces universally. This further simplifies the code by not only removing a command line option, it removes the need most of the unbounded arrays validation and several automated tests.

I also moved a bit of code that modifies the srgInfo out of the UnboundedArraysValidator, it makes more sense to have that in the client code.

Most of the changes to the automated tests should be pretty self-explanator. Some cases deserve a bit of explanation:
- unbounded-arrays-srg-layout.py was removed because this functionality is better-covered by the new srg-layouts-multiple-unbounded-arrays.azsl and verifyPackingUnboundedSpillSpaces().
- UnboundedArrays.azsl was removed because this functionality is better-covered by the new srg-layouts-multiple-unbounded-arrays.azsl and corresponding .txt pattern file.
- AzError/unbounded-arrays4.azsl was removed because there are no more register collision errors.

Signed-off-by: santorac <55155825+santorac@users.noreply.github.com>